Backport some changes in the arm assembler code as explained in this commit:
From: Russell King <rmk+kernel@arm.linux.org.uk>
Date: Mon, 19 Apr 2010 10:15:03 +0100
Subject: [PATCH 1/1] ARM: fix build error in arch/arm/kernel/process.c

/tmp/ccJ3ssZW.s: Assembler messages:
/tmp/ccJ3ssZW.s:1952: Error: can't resolve `.text' {.text section} - `.LFB1077'

This is caused because:

  .section .data
  .section .text
  .section .text
  .previous

does not return us to the .text section, but the .data section; this
makes use of .previous dangerous if the ordering of previous sections
is not known.

Fix up the other users of .previous; .pushsection and .popsection are
a safer pairing to use than .section and .previous.